+++ wagtailgeowidget/static/wagtailgeowidget/js/geocoder-field-telepath.js
@@ -19,6 +19,8 @@
         var fieldByGeocoder = {
             nominatim: NominatimGeocoderField,
             google_maps: GoogleMapsGeocoderField,
+            google_maps_places: GoogleMapsGeocoderPlacesField,
+            google_maps_places_new: GoogleMapsGeocoderPlacesNewField,
             mapbox: MapboxGeocoderField,
         };