+++ wagtailgeowidget/static/wagtailgeowidget/js/leaflet-field-telepath.js
@@ -1,7 +1,14 @@
 // This file must follow ES5
 (function () {
-    function LeafletFieldAdapter(html, _id, options) {
+    function LeafletFieldAdapter(html, options) {
         this.html = html;
+
+        // In Wagtail < 7.1 argument 2 was id
+        // TODO: Remove when Wagtail 6 is EOL
+        if (arguments[1] === "__ID__") {
+            options = arguments[2]
+        }
+
         this.options = options || {};
     }