+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Fixed
 - Replace postgres 13 to Wagtail compatible 15 in local development (@marteinn)
 - Replace deprecated google.maps.Marker with google.maps.marker.AdvancedMarkerElement (@marteinn)
+- Fix telepath path deprecation warning from Wagtail 7.1 (@marteinn)
 
 ### Removed
 - Drop support for EOL Wagtail 5.2 (@marteinn)